+++ tests/maxo/routing/conftest.py
@@ -3,7 +3,24 @@ from typing import Any
 import pytest
 
 from maxo import Bot, Ctx
+from maxo.routing.filters import BaseFilter
+from maxo.types import MessageCreated
 from tests.mocks import MockBot
+
+
+class WritingFilter(BaseFilter[MessageCreated]):
+    """Проходящий фильтр, который пишет в ctx."""
+
+    async def __call__(self, update: MessageCreated, ctx: Ctx) -> bool:
+        ctx["command"] = "start"
+        return True
+
+
+class FalseFilter(BaseFilter[MessageCreated]):
+    """Фильтр, который всегда не проходит."""
+
+    async def __call__(self, update: MessageCreated, ctx: Ctx) -> bool:
+        return False
 
 
 @pytest.fixture
